ajout MyException gestion artist sans nom gestion album sans titre

+++ src/main/java/com/myaudiolibrary/web/controller/ArtistController.java
@@ -1,6 +1,7 @@
 package com.myaudiolibrary.web.controller;
 
 
+import com.myaudiolibrary.web.exception.MyException;
 import com.myaudiolibrary.web.model.Artist;
 import com.myaudiolibrary.web.repository.ArtistRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -41,8 +42,8 @@ public class ArtistController
         Optional<Artist> artistOptional =  artistRepository.findById(id);
         if (artistOptional.isEmpty())
         {
-            System.out.println("lancer une exception");
-            throw new EntityNotFoundException("artist non trouvé");
+//            System.out.println("lancer une exception");
+            throw new EntityNotFoundException("L'artiste id "+id+" n'existe pas");
         }
 
         Artist artist = artistOptional.get();
@@ -67,7 +68,7 @@ public class ArtistController
                     @RequestParam (defaultValue = "ASC")String sortDirection
             )
     {
-        System.out.println("liste de tous les artistes");
+//        System.out.println("liste de tous les artistes");
         if (page<0)
         {
             throw new IllegalArgumentException("la valeur de page ne peut pas etre négative");
@@ -77,6 +78,11 @@ public class ArtistController
         PageRequest pageRequest = PageRequest.of(page,size, Sort.Direction.fromString(sortDirection),sortProperty);
 
         Page<Artist> artistPage = artistRepository.findAll(pageRequest);
+
+        if(page>=artistPage.getTotalPages()&&page!=0)
+        {
+            throw new EntityNotFoundException("la page demandée est supérieur aux nombre total de pages");
+        }
 
         model.put("listeArtists",artistPage);
 
@@ -100,7 +106,7 @@ public class ArtistController
                     @RequestParam (defaultValue = "ASC")String sortDirection
             )
     {
-        System.out.println("liste des artistes filtré par nom");
+//        System.out.println("liste des artistes filtré par nom");
         if (page<0)
         {
             throw new IllegalArgumentException("la valeur de page ne peut pas etre négative");
@@ -110,9 +116,17 @@ public class ArtistController
 
         Page<Artist> artistPage = artistRepository.findArtistsByNameContains(name,pageRequest);
 
+        if(page>=artistPage.getTotalPages()&&page!=0)
+        {
+            throw new EntityNotFoundException("la page demandée est supérieur aux nombre total de pages");
+        }
+
         model.put("listeArtists",artistPage);
 
-        model.put("typeRecherche",name);
+        if (!name.equals(""))
+        {
+            model.put("typeRecherche",name);
+        }
 
         return "listeArtists";
     }
@@ -129,7 +143,7 @@ public class ArtistController
     @RequestMapping(method = RequestMethod.GET, value = "/new")
     public String newArtist(final ModelMap model)
     {
-        System.out.println("on passe a la creation d'artist");
+//        System.out.println("on passe a la creation d'artist");
         Artist artist = new Artist();
 
         model.put("artist",artist);
@@ -162,6 +176,10 @@ public class ArtistController
         {
             throw new NonUniqueResultException("l'artiste "+artist.getName()+" existe déja" );
         }
+        if(artist.getName().equals(""))
+        {
+            throw  new MyException("le champ Nom ne doit pas être vide");
+        }
 
         artistRepository.save(artist);
 
@@ -188,10 +206,19 @@ public class ArtistController
         System.out.println("on met a jour l'artist ");
         System.out.println("nom de l'artist "+artist.getName());
 
+        if (!artistRepository.existsById(id))
+        {
+            throw new EntityNotFoundException("Requète impossible : l'artiste d'id "+id+" n'éxiste pas");
+        }
 
         if(artistRepository.existsByName(artist.getName()))
         {
             throw new NonUniqueResultException("un artiste avec ce nom existe déja !");
+        }
+
+        if(artist.getName().equals(""))
+        {
+            throw  new MyException("le champ Nom ne doit pas être vide");
         }
 
         Optional<Artist> artistOptional = artistRepository.findById(id);
@@ -221,7 +248,10 @@ public class ArtistController
     {
         System.out.println("delete celui la");
 
-        //erreur
+        if(!artistRepository.existsById(id))
+        {
+            throw new EntityNotFoundException("Suppresion impossible : l'artiste d'id "+id+" n'éxiste pas");
+        }
 
         artistRepository.deleteById(id);
